# doc/api.yaml
openapi: 3.0.0
info:
  title: WASAPhoto
  description: |
    API for WASAPhoto social Project.
    Each user will be presented with a stream of photos (images) in reverse chronological order, with
    information about when each photo was uploaded (date and time) and how many likes and comments
    it has. The stream is composed by photos from “following” (other users that the user follows). Users
    can place (and later remove) a “like” to photos from other users. Also, users can add comments to any
    image (even those uploaded by themself). Only authors can remove their comments.
    Users can ban other users. If user Alice bans user Eve, Eve won’t be able to see any information about
    Alice. Alice can decide to remove the ban at any moment.
    Users will have their profiles. The personal profile page for the user shows: the user’s photos (in reverse
    chronological order), how many photos have been uploaded, and the user’s followers and following.
    Users can change their usernames, upload photos, remove photos, and follow/unfollow other users.
    Removal of an image will also remove likes and comments.
    A user can search other user profiles via username.
    A user can log in just by specifying the username.
  version: 1.0.0
servers: 
  - url: 'http://localhost:3000'
...
paths:
...
  /users:
    get:
      security:
      - bearerAuth : []
      tags: ["search"]
      summary: Search user profile by username
      operationId: searchUser
      description: search other user by Username and get their profile
      parameters:
        - $ref: '#/components/parameters/username'
      responses:
        "200":
          $ref: '#/components/responses/UserProfile'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          $ref: '#/components/responses/BannedUser'
        '404':
          $ref: '#/components/responses/UserNotFound'
  /users/{userId}/profile:
    parameters:
    - $ref: '#/components/parameters/userId'
    get:
      security:
      - bearerAuth : []
      tags: ["user"]
      description: allows to get the profile of the logged user
      summary: Get user profile
      operationId: getUserProfile
      responses:
        "200":
          description: successfully return user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              examples:
                example1:
                  value:
                    userId: "123"
                    name: "Maria"
                    followerCount: 10
                    followingCount: 5
                    photosCount: 20
                    bannedUser: ["456", "789"]
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}/profile/edit:
    parameters:
      - $ref: '#/components/parameters/userId'
    put:
      security:
      - bearerAuth : []
      tags: ["user"]
      description: allows to set a new username
      summary: Set a new Username
      operationId: SetMyUserName
      requestBody:
        description: new username
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfilePrototype'
        required: true
      responses:
        "201":
          description: username updated
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfilePrototype'
              examples:
                example1:
                  value:
                   username: "Alessandro"
        "401":
          $ref: '#/components/responses/UnauthorizedError'

#-------stream of Photos-------#
  /users/{userId}/stream:
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      security:
      - bearerAuth : []
      tags: ['stream']
      summary: Get my photos stream
      description: go to my personal stream of photos
      operationId: getMyStream
      responses:
        "200":
          description: Successfully show my photo stream
          content:
            application/json:
              schema:
                description: |
                  Shows the photos of the account followed by the user logged in reverse chronological order
                type: object
                properties:
                  photos:
                    description: List of photos
                    type: array
                    items:
                      $ref: '#/components/schemas/Photo'
                    minItems: 1
                    maxItems: 100
              examples:
                example1:
                  value:
                    photos:
                      - photoId: "abc123"
                        url: "https://example.com/Users/{userId}/photos/{photosId}"
                        timestamp: "2024-05-16T17:24:49.408Z"
                        likes: ["1","2324","44"]
                        comments: ["ciao come stai?"]
        "401":
          $ref: '#/components/responses/UnauthorizedError'

#-------Photos-------#
...
  /users/{userId}/follows:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      security:
      - bearerAuth : []
      tags: ["user"]
      description: get the list of follows
      summary: retrieve list of follows for the user
      operationId: getUserfollows
      responses:
        "200": 
          description: successfully retrieve follows list
          content:
            application/json: 
              schema:
                description: list of follows
                type: object
                properties:
                  follows:
                    description: list of follows
                    type: array
                    minItems: 0
                    maxItems: 1000
                    items:
                      type: string
              examples:
                example:
                  value:
                    follows:
                      - "123"
                      - "1234"
                      - "40"
        "401": 
          $ref: "#/components/responses/UnauthorizedError"
        "404": 
          $ref: "#/components/responses/UserNotFound"
  /users/{userId}/follows/{followedId}:
    parameters:
      - $ref: '#/components/parameters/followedId'
      - $ref: '#/components/parameters/userId'
    post:
      security:
      - bearerAuth : []
      tags: ["follows"]
      description: allows to follows other accounts
      summary: follow another account
      operationId: followUser
      responses:
        "201":
          $ref: '#/components/responses/FollowUser'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags: ["follows"]
      description: allows to unfollows other accounts
      summary: another account
      operationId: unfollowUser
      responses:
        "200":
          $ref: '#/components/responses/UnfollowUser'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}/followers:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      security:
      - bearerAuth : []
      tags: ["user"]
      description: get the list of followers
      summary: retrieve list of followers for the user
      operationId: getUserfollowers
      responses:
        "200": 
          description: successfully retrieve followers list
          content:
            application/json: 
              schema:
                description: list of followers
                type: object
                properties:
                  followers:
                    description: list of followers
                    type: array
                    items:
                      type: string
                    minItems: 0
                    maxItems: 1000
              examples:
                example:
                  value:
                    followers:
                      - "123"
                      - "1234"
                      - "40"
        "401": 
          $ref: "#/components/responses/UnauthorizedError"
        "404": 
          $ref: "#/components/responses/UserNotFound"
          
#-------ban-------#
  /users/{userId}/bans/{bannedId}:
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/bannedId'
    post:
      security:
      - bearerAuth : []
      tags: ["bans"]
      description: allows to ban other accounts
      summary: ban another account
      operationId: banUser
      responses:
        "201":
          $ref: '#/components/responses/BanUser'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      security:
      - bearerAuth : []
      tags: ["bans"]
      description: allows to ban other accounts
      summary: unban another account
      operationId: unbanUser
      responses:
        "200":
          $ref: '#/components/responses/UnbanUser'
        "401":
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: string
      description: User ID used as bearer token
  responses:
...
    FollowUser: 
      description: Follow another User
      content:
        application/json:
          schema:
            description: Follow another user
            type: object
            properties:
              followsId:
                description: followed User
                type: string
                example: "1"
              userId:
                description: user to issue the follow
                type: string
                example: "123"
    UnfollowUser: 
      description: Unfollow another User
      content:
        application/json:
          schema:
            description: Unfollow another user
            type: object
            properties:
              followsId:
                description: Unfollowed User
                type: string
                example: "1"
              userId:
                description: user to issue the unfollow
                type: string
                example: "123"
    BanUser: 
      description: Ban another User
      content:
        application/json:
          schema:
            description: ban another user
            type: object
            properties:
              bannedId:
                description: bannedUser
                type: string
                example: "1"
              userId:
                description: user to issue ban
                type: string
                example: "123"
    UnbanUser: 
      description: Unban another User
      content:
        application/json:
          schema:
            description: Unban another user
            type: object
            properties:
              bannedId:
                description: bannedUser
                type: string
                example: "1"
              userId:
                description: user to issue unban
                type: string
                example: "123"
    UnauthorizedError:
      description: Access identifier is missing or invalid
      content:
        application/json:
          schema:
            description: Unauthorized error
            type: object
            properties:
              message:
                description: Error message
                type: string
                example: Unauthorized
              code:
                type: integer
                example: 401
                description: HTTP status code
    BannedUser:
      description: Forbidden, you are banned from viewing this profile
      content:
        application/json:
          schema:
            description: Banned user error
            type: object
            properties:
              message:
                description: Error message
                type: string
                example: You are banned from viewing this profile
              code:
                type: integer
                example: 403
                description: HTTP status code
    UserNotFound:
      description: A user with the specified ID was not found.
      content:
        application/json:
          schema:
            description: User not found error
            type: object
            properties:
              message:
                description: Error message
                type: string
                example: User not found
              code:
                type: integer
                example: 404
                description: HTTP status code
...
    UserProfile:
      description: Successfully return user profile
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserProfile'
          examples:
            example1:
              value:
                userId: "123"
                name: "Maria"
                followerCount: 10
                followingCount: 5
                photosCount: 20
                bannedUser: ["456", "789"]
  parameters:
    username:
      name: username
      in: query
      required: true
      description: Username of the user to search
      schema:
        description: Username of the user to search
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
    userId:
      name: userId
      in: path
      required: true
      description: ID of the user
      schema:
        description: ID of the user
        type: string    
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
...
    bannedId:
      name: bannedId
      in: path
      required: true
      description: ID of the user to ban/unban
      schema:
        description: ID of the user to ban/unban
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
    followedId:
      name: followedId
      in: path
      required: true
      description: ID of the user to follow/unfollow
      schema:
        description: ID of the user to follow/unfollow
        type: string
        pattern: '^.*?$'
        minLength: 1
        maxLength: 20
...
  schemas:
    UserProfilePrototype:
      description: User details
      type: object
      properties:
        username:
          type: string
          example: Maria
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          description: The name of the user
...
    Photo:
      description: Photo details
      type: object
      properties:
        photoId:
          type: string
          description: The unique identifier of the photo
        url:
          type: string
          format: url
          description: The URL where the photo is stored in the database
        timestamp:
          type: string
          format: date-time
          description: The date and time when the photo was uploaded
        likes:
          type: array
          minItems: 0
          maxItems: 100
          items:
            type: string
            description: IDs of the user who liked the photo
          description: user who liked the photo
        comments:
          type: array
          minItems: 0
          maxItems: 100
          items:
            type: string
            description: text of the comment
          description: list of comments
...
    UserProfile:
      description: User profile
      type: object
      properties:
        followerCount:
          type: integer
          description: The number of followers
        followers:
          type: array
          minItems: 0
          maxItems: 1000
          items:
            type: string
            description: follower user ID
          description: list of followers User IDs
        followingCount:
          type: integer
          description: The number of users followed
        follows:
          type: array
          minItems: 0
          maxItems: 1000
          items:
            type: string
            description: follow user ID
          description: list of follows User IDs
        photosCount:
          type: integer
          description: The number of photos uploaded
        bannedUser:
          type: array
          minItems: 0
          maxItems: 100
          items:
            type: string
            description: banned user ID
          description: The list of banned users
security:
  - bearerAuth: []